requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
Django==4.1.3
djangorestframework==3.14.0
geographiclib==1.52
geopy==2.2.0
idna==3.4
MarkupSafe==2.1.1
passlib==1.7.4
pycparser==2.21
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.12
Werkzeug==2.2.2
